better error logging

+++ src/vector.c
@@ -13,7 +13,7 @@ int vec_expand_(char **data, int *length, int *capacity, int memsz)
     {
         void *ptr;
         int n = (*capacity == 0) ? 1 : *capacity << 1;
-        ptr = realloc(*data, n * memsz);
+        ptr = realloc(*data, (long)n * memsz);
         if (ptr == NULL)
             return -1;
         *data = ptr;
@@ -27,7 +27,7 @@ int vec_reserve_(char **data, int *length, int *capacity, int memsz, int n)
     (void)length;
     if (n > *capacity)
     {
-        void *ptr = realloc(*data, n * memsz);
+        void *ptr = realloc(*data, (long)n * memsz);
         if (ptr == NULL)
             return -1;
         *data = ptr;
@@ -60,7 +60,7 @@ int vec_compact_(char **data, int *length, int *capacity, int memsz)
     {
         void *ptr;
         int n = *length;
-        ptr = realloc(*data, n * memsz);
+        ptr = realloc(*data, (long)n * memsz);
         if (ptr == NULL)
             return -1;
         *capacity = n;
@@ -77,7 +77,7 @@ int vec_insert_(char **data, int *length, int *capacity, int memsz,
         return err;
     memmove(*data + (idx + 1) * memsz,
             *data + idx * memsz,
-            (*length - idx) * memsz);
+            (long)(*length - idx) * memsz);
     return 0;
 }
 
@@ -87,7 +87,7 @@ void vec_splice_(char **data, int *length, int *capacity, int memsz,
     (void)capacity;
     memmove(*data + start * memsz,
             *data + (start + count) * memsz,
-            (*length - start - count) * memsz);
+            (long)(*length - start - count) * memsz);
 }
 
 void vec_swapsplice_(char **data, int *length, int *capacity, int memsz,
@@ -96,7 +96,7 @@ void vec_swapsplice_(char **data, int *length, int *capacity, int memsz,
     (void)capacity;
     memmove(*data + start * memsz,
             *data + (*length - count) * memsz,
-            count * memsz);
+            (long)count * memsz);
 }
 
 void vec_swap_(char **data, int *length, int *capacity, int memsz,
